Revert "Merge branch 'master' into Fix-validation-reset-password"
This reverts commit d02a5794a84bf82dbdca12d32424b89b50d5d919, reversing
changes made to 45be5748d3055918971fa3aa348c7e1fa93e1dae.

+++ database/migrations/2023_10_10_223422_create_pagos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('pagos', function (Blueprint $table) {
             $table->id();
             $table->date('fechapago');
-            $table->float('cantidadpago');
+            $table->integer('cantidadpago');
             $table->boolean('estadopago');
             $table->unsignedBigInteger('facturas_id');
             $table->foreign('facturas_id')->references('id')->on('facturas');
